+++ crates/azure_types/src/storage_account_blob_container_name.rs
@@ -90,6 +90,13 @@ impl FromStr for StorageAccountBlobContainerName {
         StorageAccountBlobContainerName::try_new(s)
     }
 }
+impl TryFrom<&str> for StorageAccountBlobContainerName {
+    type Error = eyre::Error;
+
+    fn try_from(value: &str) -> Result<Self, Self::Error> {
+        StorageAccountBlobContainerName::try_new(value)
+    }
+}
 impl std::fmt::Display for StorageAccountBlobContainerName {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         f.write_str(&self.inner)